+++ wltp/experiment.py
@@ -147,7 +147,9 @@ class Experiment(object):
         f_inertial = mdl["f_inertial"]
 
         v_max_rec = vmax.calc_v_max(wot, gear_ratios, f0, f1, f2, 1 - f_inertial)
-        v_max = v_max_rec.v_max
+        mdl["v_max"] = v_max = v_max_rec.v_max
+        mdl["n_v_max"] = v_max_rec.n_v_max
+        mdl["g_v_max"] = v_max_rec.g_v_max
 
         p_m_ratio = 1000 * p_rated / unladen_mass
         mdl["pmr"] = p_m_ratio