+++ CHANGES.rst
@@ -42,12 +42,15 @@ Known deficiencies
 Questions to Heinz
 ==================
 * VMax in `F new vehicle.form.vbs <https://github.com/JRCSTU/wltp/blob/master/Notebooks/WLTP_GS_AccessDB-sources/F%20new%20vehicle.form.vbs>`_:
+    * Is this the `v_max` used for class 3a/b decision?
     * L3358-L3360: is this rounding needed because of
       accumulation of rounding errors?
     * L2835:
     * Why some times down to ng-3 others ng-2, etc?  Why not scan fro top for max-v?
-      Is it pobbiel a lower gear to have lower v_max and next lower to have v_max high again??
+      Is it possible a lower gear to have lower v_max and next lower to have v_max high again??
       is there a 3-geared car with v_max@gear-1??
+    * In Annex-2.g, `v_max3` is actually :math:`(n/v)(ng_{\bold{vmax}}) \times V_{max,vehicle}`,
+      correct?
 
 
 .. _todos-list: